Extract scores export tree assembly into export_materialize.
Keep export_precedence limited to classification and payload resolution
so materialization stays one layer above pure precedence logic.
